Rename forder for constants. Edit error message.
