+++ src/Applicanty.UI/ClientApp/constants/enum-names.ts
