+++ src/Applicanty.UI/ClientApp/constants/status-commands.ts
