+++ src/Applicanty.UI/ClientApp/modules/vacancies/components/vacancy-page/vacancy-page.component.scss
@@ -41,5 +41,4 @@ table {
 .btn-notification{
     position:relative;
     overflow:unset;
-    
 }
